requirements.txt
piexif>=1.1.3
Pillow>=7.0.0
